+++ skorch/tests/callbacks/test_logging.py
@@ -14,6 +14,7 @@ from skorch.tests.conftest import neptune_installed
 from skorch.tests.conftest import sacred_installed
 from skorch.tests.conftest import wandb_installed
 from skorch.tests.conftest import tensorboard_installed
+from skorch.tests.conftest import mlflow_installed
 
 
 @pytest.mark.skipif(
@@ -941,3 +942,185 @@ class TestTensorBoard:
 
         # is not empty
         assert os.listdir(path)
+
+
+@pytest.mark.skipif(
+    not mlflow_installed, reason='mlflow is not installed')
+class TestMLflowLogger:
+    @pytest.fixture
+    def net_cls(self):
+        from skorch import NeuralNetClassifier
+        return NeuralNetClassifier
+
+    @pytest.fixture
+    def logger_cls(self):
+        from skorch.callbacks import MlflowLogger
+        return MlflowLogger
+
+    @pytest.fixture
+    def data(self, classifier_data):
+        X, y = classifier_data
+        X, y = X[:40], y[:40]
+        return X, y
+
+    @pytest.fixture
+    def mock_run(self):
+        from mlflow.entities import Run
+        return Mock(Run)
+
+    @pytest.fixture
+    def mock_client(self):
+        from mlflow.tracking import MlflowClient
+        return Mock(MlflowClient)
+
+    @pytest.fixture
+    def logger_mock_cls(self, logger_cls, mock_run, mock_client):
+        return partial(logger_cls, mock_run, mock_client)
+
+    @pytest.fixture
+    def net_builder_cls(self, net_cls, classifier_module, data):
+        def builder(*args, **kwargs):
+            return net_cls(classifier_module, *args, **kwargs).fit(*data)
+        return builder
+
+    @pytest.fixture
+    def net_fitted(self, logger_mock_cls, net_builder_cls):
+        return net_builder_cls(callbacks=[logger_mock_cls()], max_epochs=3)
+
+    def test_run_default(self, monkeypatch, logger_cls, mock_run, mock_client):
+        import mlflow
+        mock_active_run = Mock(mlflow.active_run, return_value=mock_run)
+        monkeypatch.setattr(mlflow, 'active_run', mock_active_run)
+        logger = logger_cls(client=mock_client).initialize()
+        assert mock_active_run.called
+        assert logger.run_ == mock_run
+
+    def test_client_default(self, monkeypatch, logger_cls, mock_run, mock_client):
+        import mlflow.tracking
+        monkeypatch.setattr(mlflow.tracking, 'MlflowClient', mock_client)
+        logger = logger_cls(run=mock_run).initialize()
+        assert mock_client.called
+        assert logger.client_ == mock_client()
+
+    def test_keys_from_history_logged(self, net_fitted, mock_client):
+        assert mock_client.log_metric.call_count == 3 * 4
+        keys = {call_args[0][1] for call_args in mock_client.log_metric.call_args_list}
+        expected = {'dur', 'train_loss', 'valid_loss', 'valid_acc'}
+        assert keys == expected
+
+    def test_ignore_keys(self, logger_mock_cls, net_builder_cls):
+        # ignore 'dur' and 'valid_loss', 'unknown' doesn't exist but
+        # this should not cause a problem
+        logger = logger_mock_cls(keys_ignored=['dur', 'valid_loss', 'unknown'])
+        net_builder_cls(callbacks=[logger], max_epochs=3)
+        keys = {
+            call_args[0][1]
+            for call_args in logger.client_.log_metric.call_args_list
+        }
+        expected = {'train_loss', 'valid_acc'}
+        assert keys == expected
+
+    def test_keys_ignored_is_string(self, logger_mock_cls):
+        logger = logger_mock_cls(keys_ignored='a-key').initialize()
+        expected = {'a-key', 'batches'}
+        assert logger.keys_ignored_ == expected
+
+    @pytest.mark.parametrize(
+        'log_on_batch_end, log_on_epoch_end, batch_suffix, epoch_suffix',
+        [(False, False, '', ''),
+         (True, False, '', ''),
+         (False, True, '', ''),
+         (True, True, '_batch', '_epoch')]
+    )
+    def test_epoch_batch_suffixes_defaults(
+            self,
+            logger_mock_cls,
+            log_on_batch_end,
+            log_on_epoch_end,
+            batch_suffix,
+            epoch_suffix,
+    ):
+        logger = logger_mock_cls(
+            log_on_batch_end=log_on_batch_end,
+            log_on_epoch_end=log_on_epoch_end
+        ).initialize()
+        assert logger.batch_suffix_ == batch_suffix
+        assert logger.epoch_suffix_ == epoch_suffix
+
+    @pytest.mark.parametrize('batch_suffix', ['', '_foo'])
+    @pytest.mark.parametrize('epoch_suffix', ['', '_bar'])
+    def test_epoch_batch_custom_suffix(
+            self,
+            logger_mock_cls,
+            batch_suffix,
+            epoch_suffix,
+    ):
+        logger = logger_mock_cls(
+            log_on_batch_end=True,
+            log_on_epoch_end=True,
+            batch_suffix=batch_suffix,
+            epoch_suffix=epoch_suffix,
+        ).initialize()
+        assert logger.batch_suffix_ == batch_suffix
+        assert logger.epoch_suffix_ == epoch_suffix
+
+    def test_dont_log_epoch_metrics(self, logger_mock_cls, net_builder_cls):
+        logger = logger_mock_cls(
+            log_on_batch_end=True,
+            log_on_epoch_end=False,
+            batch_suffix='_batch',
+            epoch_suffix='_epoch',
+        )
+        net_builder_cls(batch_size=10, callbacks=[logger], max_epochs=3)
+        assert all(
+            call[0][1].endswith('_batch')
+            for call in logger.client_.log_metric.call_args_list
+        )
+
+    def test_log_epochs_with_step(self, net_fitted, mock_client):
+        expected = [x for x in range(1, 4) for _ in range(4)]
+        actual = [call[1].get('step') for call in mock_client.log_metric.call_args_list]
+        assert expected == actual
+
+    def test_log_batch_with_step(self, logger_mock_cls, net_builder_cls):
+        logger = logger_mock_cls(log_on_batch_end=True, log_on_epoch_end=False)
+        net_builder_cls(batch_size=10, callbacks=[logger], max_epochs=4)
+        expected = [x for x in range(1, 21) for _ in range(2)]
+        actual = [
+            call[1].get('step')
+            for call in logger.client_.log_metric.call_args_list
+        ]
+        assert expected == actual
+
+    def test_artifact_filenames(self, net_fitted, mock_client):
+        keys = {call_args[0][1].name
+                for call_args in mock_client.log_artifact.call_args_list}
+        expected = {'params.pth', 'optimizer.pth', 'criterion.pth', 'history.json'}
+        assert keys == expected
+
+    def test_artifact_in_temporary_directory(self, net_fitted, mock_client):
+        for call_args in mock_client.log_artifact.call_args_list:
+            assert str(call_args[0][1]).startswith('/tmp')
+
+    def test_dont_create_artifact(self, net_builder_cls, logger_mock_cls):
+        logger = logger_mock_cls(create_artifact=False)
+        net_builder_cls(callbacks=[logger], max_epochs=3)
+        assert not logger.client_.log_artifact.called
+
+    def test_run_terminated_automatically(self, net_fitted, mock_client):
+        assert mock_client.set_terminated.call_count == 1
+
+    def test_run_not_closed(self, logger_mock_cls, mock_client, net_builder_cls):
+        logger = logger_mock_cls(terminate_after_train=False)
+        net_builder_cls(callbacks=[logger], max_epochs=2)
+        assert logger.client_.set_terminated.call_count == 0
+
+    def test_fit_with_real_run_and_client(self, tmp_path, logger_cls, net_builder_cls):
+        from mlflow.tracking import MlflowClient
+        client = MlflowClient(tracking_uri=tmp_path.as_uri())
+        experiment_name = 'foo'
+        experiment_id = client.create_experiment(experiment_name)
+        run = client.create_run(experiment_id)
+        logger = logger_cls(run, client, create_artifact=False)
+        net_builder_cls(callbacks=[logger], max_epochs=3)
+        assert os.listdir(tmp_path)